Move FlatViewList to org.apache.pinot.common.collections
